refactor: reorganize test files to match new src folder structure
- Move launchql/ tests to core/ (LaunchQL class tests)
- Move resolve/ tests to resolution/ (dependency resolution tests)
- Move utils/ tests to workspace/ (workspace utility tests)
- Move deploy/ tests to projects/ (project deployment tests)
- Move migrate/ tests to migration/ (migration system tests)
- Update all import paths to match new src organization
- Maintain test functionality while improving logical organization
