+++ packages/core/__tests__/core/deps-w-tags.test.ts
