+++ packages/core/__tests__/core/deps.test.ts
