+++ packages/core/__tests__/migration/cross-project.test.ts
