+++ packages/core/__tests__/modules/mods.install.test.ts
@@ -1,6 +1,6 @@
 import fs from 'fs';
 import path from 'path';
-import { LaunchQLProject } from '../../src/class/launchql';
+import { LaunchQLProject } from '../../src/core/class/launchql';
 import { TestFixture } from '../../test-utils';
 import * as glob from 'glob';
