+++ packages/core/__tests__/modules/mods.publish.test.ts
@@ -1,7 +1,7 @@
 import fs from 'fs';
 import path from 'path';
 import * as glob from 'glob';
-import { LaunchQLProject } from '../../src/class/launchql';
+import { LaunchQLProject } from '../../src/core/class/launchql';
 import { TestFixture } from '../../test-utils';
 
 let fixture: TestFixture;
