+++ packages/core/__tests__/projects/basic-deployment.test.ts
