+++ packages/core/__tests__/resolution/resolve.revert.test.ts
@@ -1,5 +1,5 @@
 import { TestFixture } from '../../test-utils';
-import { resolve } from '../../src/resolve';
+import { resolve } from '../../src/resolution/resolve';
 
 let fixture: TestFixture;
